+++ scrapers/mangabox/mangabox.go
@@ -45,8 +45,8 @@ func Mangabox(name, baseUrl, searchPath, dateLayout, dateSelector string) *scrap
 		// TODO: Parse "Vol. #"?
 		VolumeExtractor: &scraper.VolumeExtractor{
 			Selector: "body",
-			Number: func(selection *goquery.Selection) int {
-				return 1
+			Number: func(selection *goquery.Selection) float32 {
+				return 1.0
 			},
 		},
 		ChapterExtractor: &scraper.ChapterExtractor{
