+++ scrapers/mangapill/mangapill.go
@@ -57,8 +57,8 @@ var Config = &scraper.Configuration{
 	VolumeExtractor: &scraper.VolumeExtractor{
 		// selector that points to only 1 element ("Chapters" header)
 		Selector: "body > div.container > div.border.border-border.rounded > div.p-3.border-b.border-border > div.flex.flex-col.md\\:flex-row.md\\:items-center.md\\:justify-between",
-		Number: func(selection *goquery.Selection) int {
-			return 1
+		Number: func(selection *goquery.Selection) float32 {
+			return 1.0
 		},
 	},
 	ChapterExtractor: &scraper.ChapterExtractor{
